automationsAutomator/requirements.txt
Flask
Flask-WTF
playwright
requests
python-dotenv